requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.48.0
django-heroku==0.3.1
django-oauth-toolkit==1.7.0
django-rest-framework==0.1.0
django-social-auth==0.7.28
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
ecdsa==0.17.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.0
oauth2==1.9.0.post1
oauthlib==3.2.0
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-jose==3.3.0
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.13.3
